tests: Refactor and expand testing for `sntpc` functionality
Extracted tests to a `tests` directory
