+++ sntpc/src/lib.rs
@@ -646,11 +646,8 @@ pub fn fraction_to_picoseconds(sec_fraction: u32) -> u64 {
 
 #[cfg(test)]
 mod sntpc_ntp_result_tests {
+    use crate::offset_calculate;
     use crate::types::Units;
-    use crate::{
-        Error, NtpPacket, NtpResult, RawNtpPacket, SendRequestResult, fraction_to_microseconds,
-        fraction_to_milliseconds, fraction_to_nanoseconds, fraction_to_picoseconds, offset_calculate, process_response,
-    };
 
     struct Timestamps(u64, u64, u64, u64);
     struct OffsetCalcTestCase {
@@ -681,98 +678,6 @@ mod sntpc_ntp_result_tests {
         fn t4(&self) -> u64 {
             self.timestamp.3
         }
-    }
-
-    #[test]
-    fn test_ntp_result() {
-        let result1 = NtpResult::new(0, 0, 0, 0, 1, -2);
-
-        assert_eq!(0, result1.sec());
-        assert_eq!(0, result1.sec_fraction());
-        assert_eq!(0, result1.roundtrip());
-        assert_eq!(0, result1.offset());
-        assert_eq!(1, result1.stratum());
-        assert_eq!(-2, result1.precision());
-
-        let result2 = NtpResult::new(1, 2, 3, 4, 5, -23);
-
-        assert_eq!(1, result2.sec());
-        assert_eq!(2, result2.sec_fraction());
-        assert_eq!(3, result2.roundtrip());
-        assert_eq!(4, result2.offset());
-        assert_eq!(5, result2.stratum());
-        assert_eq!(-23, result2.precision());
-
-        let result3 = NtpResult::new(u32::MAX - 1, u32::MAX, u64::MAX, i64::MAX, 1, -127);
-
-        assert_eq!(u32::MAX, result3.sec());
-        assert_eq!(0, result3.sec_fraction());
-        assert_eq!(u64::MAX, result3.roundtrip());
-        assert_eq!(i64::MAX, result3.offset());
-        assert_eq!(-127, result3.precision());
-    }
-
-    #[test]
-    fn test_ntp_fraction_overflow_result() {
-        let result = NtpResult::new(0, u32::MAX, 0, 0, 1, -19);
-        assert_eq!(1, result.sec());
-        assert_eq!(0, result.sec_fraction());
-        assert_eq!(0, result.roundtrip());
-        assert_eq!(0, result.offset());
-
-        let result = NtpResult::new(u32::MAX - 1, u32::MAX, 0, 0, 1, -17);
-        assert_eq!(u32::MAX, result.sec());
-        assert_eq!(0, result.sec_fraction());
-        assert_eq!(0, result.roundtrip());
-        assert_eq!(0, result.offset());
-    }
-
-    #[test]
-    fn test_conversion_to_ms() {
-        let result = NtpResult::new(0, u32::MAX - 1, 0, 0, 1, 0);
-        let milliseconds = fraction_to_milliseconds(result.seconds_fraction);
-        assert_eq!(999u32, milliseconds);
-
-        let result = NtpResult::new(0, 0, 0, 0, 1, 0);
-        let milliseconds = fraction_to_milliseconds(result.seconds_fraction);
-        assert_eq!(0u32, milliseconds);
-    }
-
-    #[test]
-    fn test_conversion_to_us() {
-        let result = NtpResult::new(0, u32::MAX - 1, 0, 0, 1, 0);
-        let microseconds = fraction_to_microseconds(result.seconds_fraction);
-        assert_eq!(999_999u32, microseconds);
-
-        let result = NtpResult::new(0, 0, 0, 0, 1, 0);
-        let microseconds = fraction_to_microseconds(result.seconds_fraction);
-        assert_eq!(0u32, microseconds);
-    }
-
-    #[test]
-    fn test_conversion_to_ns() {
-        let result = NtpResult::new(0, u32::MAX - 1, 0, 0, 1, 0);
-        let nanoseconds = fraction_to_nanoseconds(result.seconds_fraction);
-        assert_eq!(999_999_999u32, nanoseconds);
-
-        let result = NtpResult::new(0, 0, 0, 0, 1, 0);
-        let nanoseconds = fraction_to_nanoseconds(result.seconds_fraction);
-        assert_eq!(0u32, nanoseconds);
-    }
-
-    #[test]
-    fn test_conversion_to_ps() {
-        let result = NtpResult::new(0, u32::MAX - 1, 0, 0, 1, 0);
-        let picoseconds = fraction_to_picoseconds(result.seconds_fraction);
-        assert_eq!(999_999_999_767u64, picoseconds);
-
-        let result = NtpResult::new(0, 1, 0, 0, 1, 0);
-        let picoseconds = fraction_to_picoseconds(result.seconds_fraction);
-        assert_eq!(232u64, picoseconds);
-
-        let result = NtpResult::new(0, 0, 0, 0, 1, 0);
-        let picoseconds = fraction_to_picoseconds(result.seconds_fraction);
-        assert_eq!(0u64, picoseconds);
     }
 
     #[test]
@@ -816,244 +721,8 @@ mod sntpc_ntp_result_tests {
     }
 
     #[test]
-    fn test_kiss_of_death() {
-        let defined_codes = [
-            "ACST", "AUTH", "AUTO", "BCST", "CRYP", "DENY", "DROP", "RSTR", "INIT", "MCST", "NKEY", "RATE", "RMOT",
-            "STEP",
-        ];
-
-        for code in defined_codes {
-            let ref_id_bytes = code.as_bytes();
-            let ref_id = u32::from_be_bytes(ref_id_bytes.try_into().unwrap_or([0; 4]));
-            let req_resp = SendRequestResult {
-                originate_timestamp: 0u64,
-                version: 0u8,
-            };
-            let response = NtpPacket {
-                li_vn_mode: 0xC4,
-                stratum: 0,
-                poll: 0,
-                precision: 0,
-                root_delay: 0,
-                root_dispersion: 0,
-                ref_id,
-                ref_timestamp: 0,
-                origin_timestamp: 0,
-                recv_timestamp: 0,
-                tx_timestamp: 0,
-            };
-            let raw: RawNtpPacket = (&response).into();
-            let result = process_response(req_resp, raw, 0);
-
-            assert!(result.is_err());
-
-            match result.unwrap_err() {
-                Error::KissOfDeath(kod_code) => assert_eq!(kod_code.as_str(), code),
-                _ => unreachable!("Unexpected error code"),
-            }
-        }
-    }
-}
-
-#[cfg(all(test, feature = "std"))]
-mod sntpc_std_tests {
-    use crate::types::Units;
-
-    #[test]
     fn test_units_str_representation() {
         assert_eq!(format!("{}", Units::Milliseconds), "ms");
         assert_eq!(format!("{}", Units::Microseconds), "us");
-    }
-}
-
-#[cfg(all(test, feature = "std", feature = "sync"))]
-mod sntpc_sync_tests {
-    use crate::sync::get_time;
-    use crate::{Error, NtpContext, NtpUdpSocket, StdTimestampGen};
-    use std::net::{SocketAddr, ToSocketAddrs, UdpSocket};
-
-    struct UdpSocketWrapper {
-        socket: UdpSocket,
-    }
-
-    impl UdpSocketWrapper {
-        #[must_use]
-        fn new(socket: UdpSocket) -> Self {
-            Self { socket }
-        }
-    }
-
-    impl From<UdpSocket> for UdpSocketWrapper {
-        fn from(socket: UdpSocket) -> Self {
-            UdpSocketWrapper::new(socket)
-        }
-    }
-
-    impl NtpUdpSocket for UdpSocketWrapper {
-        async fn send_to(&self, buf: &[u8], addr: SocketAddr) -> crate::Result<usize> {
-            match self.socket.send_to(buf, addr) {
-                Ok(usize) => Ok(usize),
-                Err(_) => Err(Error::Network),
-            }
-        }
-
-        async fn recv_from(&self, buf: &mut [u8]) -> crate::Result<(usize, SocketAddr)> {
-            match self.socket.recv_from(buf) {
-                Ok((size, addr)) => Ok((size, addr)),
-                Err(_) => Err(Error::Network),
-            }
-        }
-    }
-
-    #[test]
-    fn test_ntp_request_sntpv4_supported() {
-        let context = NtpContext::new(StdTimestampGen::default());
-        let pools = [
-            "pool.ntp.org:123",
-            "time.google.com:123",
-            "time.apple.com:123",
-            "time.cloudflare.com:123",
-            "time.facebook.com:123",
-        ];
-
-        for pool in &pools {
-            let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-            socket
-                .set_read_timeout(Some(std::time::Duration::from_secs(2)))
-                .expect("Unable to set up socket timeout");
-            let socket = UdpSocketWrapper::from(socket);
-
-            for address in pool.to_socket_addrs().unwrap().filter(SocketAddr::is_ipv4) {
-                let result = get_time(address, &socket, context);
-
-                assert!(result.is_ok(), "{} is bad - {:?}", pool, result.unwrap_err());
-                assert_ne!(result.unwrap().seconds, 0);
-            }
-        }
-    }
-
-    #[test]
-    fn test_ntp_request_sntpv3_not_supported() {
-        let context = NtpContext::new(StdTimestampGen::default());
-
-        let pools = ["time.nist.gov:123", "time.windows.com:123"];
-
-        for pool in &pools {
-            let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-            socket
-                .set_read_timeout(Some(std::time::Duration::from_secs(5)))
-                .expect("Unable to set up socket timeout");
-            let socket = UdpSocketWrapper::from(socket);
-
-            for address in pool.to_socket_addrs().unwrap().filter(SocketAddr::is_ipv4) {
-                let result = get_time(address, &socket, context);
-                assert!(result.is_err(), "{pool} is ok");
-                assert_eq!(result.unwrap_err(), Error::IncorrectResponseVersion);
-            }
-        }
-    }
-
-    #[test]
-    fn test_invalid_addrs_ntp_request() {
-        let pool = "asdf.asdf.asdf:123";
-        let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-        socket
-            .set_read_timeout(Some(std::time::Duration::from_secs(5)))
-            .expect("Unable to set up socket timeout");
-
-        let result = pool.to_socket_addrs();
-        assert!(result.is_err(), "{pool} is ok");
-    }
-}
-
-#[cfg(all(test, feature = "std", feature = "sync"))]
-mod sntpc_async_tests {
-    use crate::sync::SYNC_EXECUTOR_NUMBER_OF_TASKS;
-    use crate::{Error, NtpContext, StdTimestampGen};
-    use crate::{NtpUdpSocket, get_time};
-    use miniloop::executor::Executor;
-    use std::net::{SocketAddr, ToSocketAddrs, UdpSocket};
-
-    struct UdpSocketWrapper {
-        socket: UdpSocket,
-    }
-
-    impl UdpSocketWrapper {
-        #[must_use]
-        fn new(socket: UdpSocket) -> Self {
-            Self { socket }
-        }
-    }
-
-    impl From<UdpSocket> for UdpSocketWrapper {
-        fn from(socket: UdpSocket) -> Self {
-            UdpSocketWrapper::new(socket)
-        }
-    }
-
-    impl NtpUdpSocket for UdpSocketWrapper {
-        async fn send_to(&self, buf: &[u8], addr: SocketAddr) -> crate::Result<usize> {
-            match self.socket.send_to(buf, addr) {
-                Ok(usize) => Ok(usize),
-                Err(_) => Err(Error::Network),
-            }
-        }
-
-        async fn recv_from(&self, buf: &mut [u8]) -> crate::Result<(usize, SocketAddr)> {
-            match self.socket.recv_from(buf) {
-                Ok((size, addr)) => Ok((size, addr)),
-                Err(_) => Err(Error::Network),
-            }
-        }
-    }
-
-    #[test]
-    fn test_ntp_async_request_sntpv4_supported() {
-        let context = NtpContext::new(StdTimestampGen::default());
-        let pools = [
-            "pool.ntp.org:123",
-            "time.google.com:123",
-            "time.apple.com:123",
-            "time.cloudflare.com:123",
-            "time.facebook.com:123",
-        ];
-
-        for pool in &pools {
-            let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-            socket
-                .set_read_timeout(Some(std::time::Duration::from_secs(5)))
-                .expect("Unable to set up socket timeout");
-            let socket = UdpSocketWrapper::from(socket);
-
-            for address in pool.to_socket_addrs().unwrap().filter(SocketAddr::is_ipv4) {
-                let result =
-                    Executor::<SYNC_EXECUTOR_NUMBER_OF_TASKS>::new().block_on(get_time(address, &socket, context));
-
-                assert!(result.is_ok(), "{pool} is bad - {:?}", result.unwrap_err());
-                assert_ne!(result.unwrap().seconds, 0);
-            }
-        }
-    }
-
-    #[test]
-    fn test_ntp_async_request_sntpv3_not_supported() {
-        let context = NtpContext::new(StdTimestampGen::default());
-
-        let pools = ["time.nist.gov:123", "time.windows.com:123"];
-
-        for pool in &pools {
-            let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-            socket
-                .set_read_timeout(Some(std::time::Duration::from_secs(5)))
-                .expect("Unable to set up socket timeout");
-            let socket = UdpSocketWrapper::from(socket);
-
-            for address in pool.to_socket_addrs().unwrap().filter(SocketAddr::is_ipv4) {
-                let result =
-                    Executor::<SYNC_EXECUTOR_NUMBER_OF_TASKS>::new().block_on(get_time(address, &socket, context));
-                assert!(result.is_err(), "{pool} is ok");
-                assert_eq!(result.unwrap_err(), Error::IncorrectResponseVersion);
-            }
-        }
     }
 }
